+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `load_model()` functions now auto-detects if a model is multi_inputs or not
+- Python version support was changed to 3.8, 3.9, 3.10 (other versions should still work but are not systematically tested)
 
 ## [0.3.1] - 2023-01-06